refactor: improve speed and remaining time calculate thread
adjust atomic memory order

+++ src/logic/DownloadManager.h
@@ -26,7 +26,8 @@ public:
     [[nodiscard]] std::size_t addTask(const std::string& url, const std::string& filePath,
                                       unsigned int threadNum = std::thread::hardware_concurrency());
 
-    using ProgressCallback = std::function<void(unsigned long total, unsigned long downloaded, double speed, double remainTime)>;
+    using ProgressCallback =
+            std::function<void(unsigned long total, unsigned long downloaded, unsigned long speed, double remainTime)>;
     using DownloadCompleteCallback = std::function<void()>;
     void setProgressCallback(size_t taskID, const ProgressCallback& cb);
     void setDownloadCompleteCallback(size_t taskID, const DownloadCompleteCallback& cb);
